% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_buttons.R
\name{arrowButton}
\alias{arrowButton}
\title{Arrow button}
\usage{
arrowButton(id, direction = "left", ...)
}
\arguments{
\item{id}{The input id to read value from.}

\item{direction}{The direction of arrow button.}

\item{...}{List of tag attributes.}
}
\description{
Arrow button
}
\examples{
if (interactive()) {
  arrowButton("awesome-id", direction = "left")
}
}
