% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_waiters.R
\name{dcWaiter}
\alias{dcWaiter}
\title{Title}
\usage{
dcWaiter(
  stage = c("show", "update", "hide"),
  id = NULL,
  user.name = "DCA User",
  is.landing = FALSE,
  is.stop = FALSE,
  sleep = 2,
  is.certified = TRUE,
  is.permission = TRUE,
  msg = NULL,
  spin = NULL,
  color = NULL
)
}
\arguments{
\item{sleep}{recommend to give at least 2s to let users view the page before updating
initial screen will not use _sleep}

\item{spin}{}
}
\value{

}
\description{
Title
}
