% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_empty}
\alias{drop_empty}
\title{drop_empty}
\usage{
drop_empty(x, drop = "all")
}
\arguments{
\item{x}{List of input values}

\item{drop}{Type of values to drop. Default to 'all'.}
}
\value{
clean data
}
\description{
drop_empty
}
\examples{
if (interactive()) {
  values <- c("real", "null", NULL, NA, "  ")
  drop_empty(values)
  # [1] "real"
}
}
