% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_header.R
\name{expandHeader}
\alias{expandHeader}
\title{Expand dashboardheader}
\usage{
expandHeader(.header, left = NULL)
}
\arguments{
\item{left}{The left menu in the header.}

\item{header}{\code{dashboardheader}.}
}
\description{
Expand the ui to \code{dashboardheader}.
}
\examples{
if (interactive()) {
  library(shinydashboard)
  dashboardheader() \%>\%
    expandHeader()
  )
}
}
