% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tab_names}
\alias{get_tab_names}
\title{Extract tab names}
\usage{
get_tab_names(.tab)
}
\arguments{
\item{.tab}{.tab The \code{tabItem} object.}
}
\value{
A vector of tab names.
}
\description{
Get tab names from \code{tabItem} object.
}
\examples{
\dontrun{
tab_object <- tabItems(
  tabItem(tabName = "tab1"),
  tabItem(tabName = "tab2"),
  tabItem(tabName = "tab3")
)
print(get_tab_names(tab_object))
# [1] "tab1" "tab2" "tab3"
}
}
