% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_attr_box}
\alias{insert_attr_box}
\title{Insert attribute to a shinydashboard::box}
\usage{
insert_attr_box(.box, ...)
}
\arguments{
\item{...}{The attributes added to the box}

\item{.tag}{The \code{shinydashboard::box} object.}
}
\value{
New box object after adding attributes.
}
\description{
Insert attribute to a shinydashboard::box
}
\examples{
\dontrun{
insert_attribute(box(), id = "dca")
<div class="col-sm-6">
  <div class="box" id="dca">
  <div class="box-body"></div>
  </div>
</div>
}
}
