% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_attribute}
\alias{insert_attribute}
\title{Insert attribute to a tag}
\usage{
insert_attribute(.tag, pos = 1, ...)
}
\arguments{
\item{.tag}{The tag object.}

\item{pos}{List of tag attributes. \code{pos = 0} adds attribute to \code{.tag}. \code{pos = 1} adds attribute to the first child of \code{.tag}. Default to 1.}

\item{...}{After which position to insert attributes.}
}
\value{
New tag object after adding attributes.
}
\description{
Insert attribute to a tag
}
\examples{
\dontrun{
insert_attribute(div(), id = "dca")
# <div id="dca"></div>
}
}
