% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_buttons.R
\name{mediaButton}
\alias{mediaButton}
\title{Media button}
\usage{
mediaButton(icon, link = NULL, ...)
}
\arguments{
\item{icon}{The icon object from \code{shiny::icon}}

\item{link}{The redirecting url}

\item{...}{List of tag attributes}
}
\description{
Media button
}
