% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_buttons.R
\name{paletteButton}
\alias{paletteButton}
\title{Palette button}
\usage{
paletteButton(id, color, ...)
}
\arguments{
\item{id}{The input id to read value from}

\item{color}{The color of palette}

\item{...}{List of tag attributes}
}
\description{
Palette button
}
