% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_panel.R
\name{palettePanel}
\alias{palettePanel}
\title{Title}
\usage{
palettePanel(id, header.id, parent.session, parent.input, parent.output)
}
\arguments{
\item{parent.output}{}
}
\value{

}
\description{
Title
}
