% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_buttons.R
\name{shinyButton}
\alias{shinyButton}
\title{shinyButton}
\usage{
shinyButton(id, label = NULL, ...)
}
\description{
A button with shiny effect
}
\examples{

if (interactive()) {

 shinyButton("button")

}
}
