% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\name{showcase}
\alias{showcase}
\title{Launch the gallery of dcamodules}
\usage{
showcase()
}
\description{
A gallery of all components available in dcamodules
}
\examples{
if (interactive()) {

 dcamodules::showcase()

}
}
