% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_waiters.R
\name{spin_logo}
\alias{spin_logo}
\title{Synapse Logo Spinner}
\usage{
spin_logo(logo = "synapse")
}
\arguments{
\item{logo}{name of logo}
}
\value{

}
\description{
Synapse logo spinner to use with waiter
}
