% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_switch_button.R
\name{tabSwitchUI}
\alias{tabSwitchUI}
\title{Create switch button}
\usage{
tabSwitchUI(.tab, id = NULL)
}
\arguments{
\item{.tab}{The \code{tabItem} object}

\item{id}{The input variable to read value from}
}
\description{
Create switch button
}
