% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{update_css}
\alias{update_css}
\title{Converting scss files to update 'main.min.css'}
\usage{
update_css(main.scss = NULL)
}
\arguments{
\item{main.scss}{path to main.min.css}
}
\value{
path of main.min.css
}
\description{
Converting scss files to update 'main.min.css'
}
