% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_waiters.R
\name{use_dca_waiter}
\alias{use_dca_waiter}
\title{Config waiter}
\usage{
use_dca_waiter()
}
\description{
Use to config dependencies of dca waiter
}
\examples{
if (interactive()) {
  library(shinydashboard)
  library(dcamodules)
  dashboardBody(
    use_dca_waiter()
  )
}
}
