% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var2server}
\alias{var2server}
\title{var2server}
\usage{
var2server(id, values)
}
\arguments{
\item{id}{The input variable to read value from}

\item{values}{The values need to be sent}
}
\description{
Use this trick to send variables from ui to server in module
}
