% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_to_server}
\alias{var_to_server}
\title{Send values from module UI to module Server}
\usage{
var_to_server(id, values)
}
\arguments{
\item{id}{The input variable to read value from.}

\item{values}{The values need to be sent.}
}
\value{
The hidden \code{selectInput} object.
}
\description{
Use this trick to send values from ui to server in module.
}
\examples{
\dontrun{
widgetUI <- function(id) {
  ns <- NS(id)
  items <- c("chicken", "egg")
  var_to_server(ns("items"), tab_names)
}

widgetServer <- function() {
  moduleServer(
    id,
    function(input, output, session) {
      # items now can be read in the server
      purrr::map_chr(isolate(input$items), print)
    }
  )
}
}
}
