% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_footer.R
\name{dcaFooter}
\alias{dcaFooter}
\title{Data Curator Footer}
\usage{
dcaFooter(message, media = NULL, height = "100px", ...)
}
\arguments{
\item{message}{The main text section at the bottom of footer.}

\item{media}{The top section for media links. Default to NULL.}

\item{height}{The height of footer. Default to '100px'/}

\item{...}{List of tag attributes}
}
\description{
Add footer to the bottom of \code{dashboardSidebar} or \code{dashboardBody}.
}
\examples{
if (interactive()) {
  library(shinydashboard)
  dashboardBody(
    dcaFooter("
      Copy right 2022",
      media = mediaButton("google", "https://google.com")
    )
  )
}
}
