% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_palette_panel.R
\name{palettePanel}
\alias{palettePanel}
\title{Palette panel server}
\usage{
palettePanel(id, head.id, parent.session, parent.input, parent.output)
}
\arguments{
\item{id}{The id of the output object.}

\item{head.id}{The id of head tag that used to change styles}

\item{parent.session}{Session from parent scope}

\item{parent.input}{Input from parent scope}

\item{parent.output}{Output from parent scope}
}
\description{
Palette panel server
}
