% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_palette_panel.R
\name{palettePanelUI}
\alias{palettePanelUI}
\title{Create an palette panel output element}
\usage{
palettePanelUI(id)
}
\arguments{
\item{id}{output variable to read the value from}
}
\description{
Create an palette panel output element
}
