% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_switch_button.R
\name{tabSwitch}
\alias{tabSwitch}
\title{Tab switch button server}
\usage{
tabSwitch(id, tab.id, parent.session, parent.input, parent.output)
}
\arguments{
\item{id}{The input id to read value from}

\item{tab.id}{The id of \code{tabItem} object}

\item{parent.session}{The session from parent scope}

\item{parent.input}{The input from parent scope}

\item{parent.output}{The output from parent scope}
}
\description{
Tab switch button server
}
