% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_switch_button.R
\name{tabSwitch}
\alias{tabSwitch}
\title{Tab switch button server}
\usage{
tabSwitch(id, tab.id, parent.session)
}
\arguments{
\item{id}{The input id to read value from.}

\item{tab.id}{The id of \code{sidebarMenu} object.}

\item{parent.session}{The session from parent scope.}
}
\description{
Tab switch button server
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinydashboard)
  library(dcamodules)
  ui <- dashboardPage(
    dashboardHeader(),
    dashboardSidebar(
      sidebarMenu(
        id = "tabs",
        menuItem("awesome_tab")
      )
    ),
    dashboardBody(
      tabSwitchUI(
        id = "switch",
        tabItems(tabItem("awesome_tab"))
      )
    )
  )
  server <- function(input, output, session) {
    tabSwitch("switch", "tabs", session)
  }
  shinyApp(ui, server)
}
}
